+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.2
 
 * Add htmlFor/id attribute to labels associating them to inputs
+* Changed button color palette
 
 ## 2.1.1